Almost back to spawing portals, npc's etc.   I have a bug in the create routine but it is close. Data is loading and createobj messages are sent to the client.   They are correclty formed, but have some bad data in them.

+++ Source/ACE.Entity/AceObject.cs
@@ -15,7 +15,7 @@ namespace ACE.Entity
             get { return new Position(Landblock, PosX, PosY, PosZ, QX, QY, QZ, QW); }
         }
 
-        [DbField("landblock", (int)MySqlDbType.UInt16)]
+        [DbField("landblock", (int)MySqlDbType.UInt32)]
         public uint Landblock { get; set; }
 
         [DbField("posX", (int)MySqlDbType.Float)]
